Reorganize additional test cases
There were a few test case functions that still weren't grouped into the
proper sub-module.
